backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pydantic-settings
supabase
